Move SCSI2SD cdrom.c to ZuluSCSI_cdrom.cpp for extensibility
No functional changes in this commit, reorganizing code
to allow extending CD-ROM emulation functionality.
